added another example of reporting on state
There is no need to communicate state of system from event queue.
All agents are owned by environment.
So it is trivial to implement reporting functions for concrete environment.
This solution is good enough for now. When I will start working in plugin system, I might revisit question of unification of reporting